aanpassingen: - Nederlandse code naar engels veranderd - namespacing toegevoegd - api coordinaten formatting bugfix
