+++ PollenBackend/Models/WeerData.cs
@@ -1,6 +1,9 @@
-public class WeerData
+namespace PollenBackend.Models
 {
-    public double AverageTemperature { get; set; }
-    public double AverageRain { get; set; }
-    public double AverageWind { get; set; }
+    public class WeatherData
+    {
+        public double AverageTemperature { get; set; }
+        public double AverageRain { get; set; }
+        public double AverageWind { get; set; }
+    }
 }
